Navigation: Central position of arrow for active element
The arrow had been positioned roughly in the middle of the navigation
element. However the exact position differed depending on the width of
the navigation element.

Now the position is always centered. To do that the element is moved to
50% from the left side of the navigation item. This would center the
left side of the arrow. Afterwards the arrow is move to 10 pixel left
using margin-left, which is the half of the arrow width.